fix some bugs
fix some bugs

+++ tessar-netty-server/src/main/java/vova/domain/newadd/NewAddDay.java
@@ -10,8 +10,7 @@ public class NewAddDay extends NewAdd{
     public NewAddDay() {
     }
 
-    public NewAddDay(int id, Date dateID, String cID, String gID, String sID, long newAddNum, long activeNum, long loginCount, long averageLogin, long allPlayerNum) {
+    public NewAddDay(int id, Date dateID, String cID, String gID, String sID, long newAddNum, long activeNum, long loginCount, float averageLogin, long allPlayerNum) {
         super(id, dateID, cID, gID, sID, newAddNum, activeNum, loginCount, averageLogin, allPlayerNum);
     }
-
 }
