+++ tessar-netty-server/src/main/java/vova/domain/newadd/NewAddMon.java
@@ -10,7 +10,7 @@ public class NewAddMon extends NewAdd {
     public NewAddMon() {
     }
 
-    public NewAddMon(int id, Date dateID, String cID, String gID, String sID, long newAddNum, long activeNum, long loginCount, long averageLogin, long allPlayerNum) {
+    public NewAddMon(int id, Date dateID, String cID, String gID, String sID, long newAddNum, long activeNum, long loginCount, float averageLogin, long allPlayerNum) {
         super(id, dateID, cID, gID, sID, newAddNum, activeNum, loginCount, averageLogin, allPlayerNum);
     }
 }
